==> week_3_classification/notebooks/1_data_preparation.ipynb <==
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt

df = pd.read_csv('../data/telco_customer_churn.csv')

df.head()

df.head(3).T

df.columns = df.columns.str.lower().str.replace(' ', '_')

categorical_columns = list(df.dtypes[df.dtypes == 'object'].index)

for c in categorical_columns:
    df[c] = df[c].str.lower().str.replace(' ', '_')

df.head(3).T

df.dtypes

df['totalcharges'] = pd.to_numeric(df['totalcharges'], errors='coerce')
df['totalcharges'] = df['totalcharges'].fillna(0)

df[df['totalcharges'] == '_']

df.churn

df.churn = (df.churn == 'yes').astype(int)

df

df.to_parquet('../data/data_prep.parquet')
