==> week_3_classification/notebooks/2_setting_up_validation_framework.ipynb <==
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt



from sklearn.model_selection import train_test_split

df = pd.read_parquet('../data/data_prep.parquet')

df.head().T

df_full_train, df_test = train_test_split(df, test_size=0.2, random_state=1)

len(df_full_train), len(df_test)

df_train, df_val = train_test_split(df_full_train, test_size=0.25, random_state=1)

len(df_train), len(df_val), len(df_test)

df_train = df_train.reset_index(drop=True)
df_val = df_val.reset_index(drop=True)
df_test = df_test.reset_index(drop=True)

y_train = df_train.churn.values
y_val = df_val.churn.values
y_test = df_test.churn.values

del df_train['churn']
del df_val['churn']
del df_test['churn']

df_train.to_parquet('../data/df_train.parquet')
df_val.to_parquet('../data/df_val.parquet')
df_test.to_parquet('../data/df_test.parquet')
df_full_train.to_parquet('../data/df_full_train.parquet')

np.savetxt('../data/y_train', y_train)
np.savetxt('../data/y_val', y_val)
np.savetxt('../data/y_test', y_test)
