==> week_3_classification/notebooks/3_EDA.ipynb <==
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt

from IPython.display import display

from sklearn.model_selection import train_test_split
from sklearn.metrics import mutual_info_score
from sklearn.feature_extraction import DictVectorizer

df = pd.read_parquet('../data/data_prep.parquet')

df_train = pd.read_parquet('../data/df_train.parquet')
df_val = pd.read_parquet('../data/df_val.parquet')
df_test = pd.read_parquet('../data/df_test.parquet')
df_full_train = pd.read_parquet('../data/df_full_train.parquet')

y_train = np.loadtxt('../data/y_train')
y_val = np.loadtxt('../data/y_val')
y_test = np.loadtxt('../data/y_test')


df_full_train = df_full_train.reset_index(drop=True)
df_full_train

# Null values

df_full_train.isnull().sum()

# Distribution churn variable

df_full_train.churn.value_counts()


df_full_train.churn.value_counts(normalize=True)

global_churn_rate = df_full_train.churn.mean()
round(global_churn_rate, 4)

df_full_train.dtypes

numerical_variables = ['tenure', 'monthlycharges', 'totalcharges']

list(df_full_train.columns)

categorical_variables = [
    'gender',
    'seniorcitizen',
    'partner',
    'dependents',
    'phoneservice',
    'multiplelines',
    'internetservice',
    'onlinesecurity',
    'onlinebackup',
    'deviceprotection',
    'techsupport',
    'streamingtv',
    'streamingmovies',
    'contract',
    'paperlessbilling',
    'paymentmethod',
]

df_full_train[categorical_variables].nunique()

## Feature importance: churn rate and risk ratio

### Churn Rate

churn_female = df_full_train[df_full_train.gender == 'female'].churn.mean()
churn_female

churn_male = df_full_train[df_full_train.gender == 'male'].churn.mean()
churn_male

global_churn_rate = df_full_train.churn.mean()
global_churn_rate

churn_partner = df_full_train[df_full_train.partner == 'yes'].churn.mean()
churn_partner

churn_no_partner = df_full_train[df_full_train.partner == 'no'].churn.mean()
churn_no_partner

### Risk Ratio

churn_no_partner / global_churn_rate # Si es mayor a 1 es más probable que abandonen

churn_partner / global_churn_rate

## Análisis de churn por variables categóricas

for c in categorical_variables:
    print(c)
    df_group = df_full_train.groupby(c).churn.agg(['mean', 'count'])
    df_group['diff'] = df_group['mean'] - global_churn_rate
    df_group['risk'] = df_group['mean'] / global_churn_rate
    display(df_group)
    print()
    print()
    

## Feature Importance: Mutual Information (for categorical variables)

Mutual Information is a concept from information theory, it tell us how much we can learn about one variable if we know the value of another

mutual_info_score(df_full_train.churn, df_full_train.contract)

mutual_info_score(df_full_train.churn, df_full_train.gender)

def mutual_info_churn_score(series):
    return mutual_info_score(series, df_full_train.churn)

mutual_information_scores = df_full_train[categorical_variables].apply(mutual_info_churn_score)
mutual_information_scores.sort_values(ascending=False)

## Feature Importance: Correlation (for numerical variables)

df_full_train[numerical_variables]

df_full_train[numerical_variables].corrwith(df_full_train.churn)

df_full_train[numerical_variables].corrwith(df_full_train.churn).abs()

For tenure

df_full_train[df_full_train.tenure <= 2].churn.mean()

df_full_train[(df_full_train.tenure > 2) & (df_full_train.tenure <=12)].churn.mean()

df_full_train[df_full_train.tenure > 12].churn.mean()

for monthlycharges

df_full_train[df_full_train.monthlycharges <= 20].churn.mean()

df_full_train[(df_full_train.monthlycharges > 20) & (df_full_train.monthlycharges <=50)].churn.mean()

df_full_train[df_full_train.monthlycharges > 50].churn.mean()

## One Hot Encoding - Using Scikit-Learn

Dict Vectorizer Example

df_train[['gender', 'contract']].iloc[:10].to_dict(orient='records')

dicts = df_train[['gender', 'contract']].iloc[:100].to_dict(orient='records')

dv = DictVectorizer(sparse=False)
dv.fit(dicts)

dv.get_feature_names_out()

dv.transform(dicts)

Apply Dict Vectorizer

train_dicts = df_train[categorical_variables + numerical_variables].to_dict(orient='records')

dv = DictVectorizer(sparse=False)
dv.fit(train_dicts)

dv.get_feature_names_out()

dv.transform(train_dicts)[0]

X_train = dv.fit_transform(train_dicts)
X_train[0]

val_dicts = df_val[categorical_variables + numerical_variables].to_dict(orient='records')
X_val = dv.transform(val_dicts)

np.savetxt('../data/X_train', X_train)
np.savetxt('../data/X_val', X_val)
