==> week_3_classification/notebooks/4_logistic_regression.ipynb <==
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt

from IPython.display import display

from sklearn.model_selection import train_test_split
from sklearn.metrics import mutual_info_score
from sklearn.feature_extraction import DictVectorizer
from sklearn.linear_model import LogisticRegression

df = pd.read_parquet('../data/data_prep.parquet')

df_train = pd.read_parquet('../data/df_train.parquet')
df_val = pd.read_parquet('../data/df_val.parquet')
df_test = pd.read_parquet('../data/df_test.parquet')
df_full_train = pd.read_parquet('../data/df_full_train.parquet')

y_train = np.loadtxt('../data/y_train')
y_val = np.loadtxt('../data/y_val')
y_test = np.loadtxt('../data/y_test')

X_train = np.loadtxt('../data/X_train')
X_val = np.loadtxt('../data/X_val')

## Logistic Regression

model = LogisticRegression()
model.fit(X_train, y_train)

model.intercept_[0]

model.coef_[0].round(3)

model.predict(X_train) # Hard Predictions

model.predict_proba(X_train) # Soft Predictions

model.predict_proba(X_train)[:, 1] # Soft Predictions

y_pred = model.predict_proba(X_val)[:, 1]
y_pred

churn_decision = (y_pred >= 0.5)

# Gente que probablemente (0.5 de threshold) va a churn

df_val[churn_decision].customerid

## Accuracy of the model

churn_decision.astype('int')

y_val

(y_val == churn_decision)

(y_val == churn_decision).mean()

df_pred = pd.DataFrame()
df_pred['probability'] = y_pred
df_pred['prediction'] = churn_decision.astype(int)
df_pred['actual'] = y_val
df_pred['correct_predictions'] = df_pred.prediction == df_pred.actual
df_pred

df_pred.correct_predictions.mean()

## Model Interpretation

numerical_variables = ['tenure', 'monthlycharges', 'totalcharges']
categorical_variables = [
    'gender',
    'seniorcitizen',
    'partner',
    'dependents',
    'phoneservice',
    'multiplelines',
    'internetservice',
    'onlinesecurity',
    'onlinebackup',
    'deviceprotection',
    'techsupport',
    'streamingtv',
    'streamingmovies',
    'contract',
    'paperlessbilling',
    'paymentmethod',
]

train_dicts = df_train[categorical_variables + numerical_variables].to_dict(orient='records')
dv = DictVectorizer(sparse=False)
dv.fit(train_dicts)
dv.get_feature_names_out()

# Take a look at the coefficients

dict(zip(dv.get_feature_names_out(), model.coef_[0].round(3)))

Create a smaller model

small = ['contract', 'tenure', 'monthlycharges']

df_train[small].iloc[:10].to_dict(orient='records')

dicts_train_small = df_train[small].to_dict(orient='records')
dicts_val_small = df_val[small].to_dict(orient='records')

dv_small = DictVectorizer(sparse=False)
dv_small.fit(dicts_train_small)

dv_small.get_feature_names_out()

X_train_small = dv_small.transform(dicts_train_small)

model_small = LogisticRegression()
model_small.fit(X_train_small, y_train)

w0 = model_small.intercept_[0]
w0.round(3)

w = model_small.coef_[0]
w.round(3)

dict(zip(dv_small.get_feature_names_out(), w.round(3)))

def sigmoid(z):
    return 1 / (1 + np.exp(-z))

# Example

w0 + 1*0.97 + 50*0.027 + 5*(-0.036)

sigmoid(_) # _ toma el valor de la salida inmediatamente anterior 

# Example

sigmoid(w0 + 1*0.97 + 60*0.027 + 1*(-0.036))

## Using the model

First, train a new model

dicts_full_train = df_full_train[categorical_variables + numerical_variables].to_dict(orient='records')

dv = DictVectorizer(sparse=False)
X_full_train = dv.fit_transform(dicts_full_train)

y_full_train = df_full_train.churn.values
y_full_train

model = LogisticRegression()
model.fit(X_full_train, y_full_train)

dicts_test = df_test[categorical_variables + numerical_variables].to_dict(orient='records')
X_test = dv.transform(dicts_test)

y_pred = model.predict_proba(X_test)[:, 1]

churn_decision = (y_pred >= 0.5)


(churn_decision == y_test).mean()

Use the model

customer = dicts_test[10]
customer

X_customer = dv.transform([customer])

X_customer.shape

model.predict_proba(X_customer)[0,1]

y_test[10]

customer_2 = dicts_test[-1]
customer_2

X_customer_2 = dv.transform([customer_2])

X_customer_2.shape

model.predict_proba(X_customer_2)[0,1]

y_test[-1]
